Fixes clang warning diff_match_patch.h:1618:20: comparison of constant 18446744073709551615 with expression of type 'int' is always true [-

+++ diff_match_patch.h
@@ -1609,7 +1609,7 @@ class diff_match_patch {
     // Highest score beyond which we give up.
     double score_threshold = Match_Threshold;
     // Is there a nearby exact match? (speedup)
-    int best_loc = text.find(pattern, loc);
+    size_t best_loc = text.find(pattern, loc);
     if (best_loc != string_t::npos) {
       score_threshold = std::min(match_bitapScore(0, best_loc, loc, pattern),
           score_threshold);
@@ -1678,7 +1678,7 @@ class diff_match_patch {
             best_loc = j - 1;
             if (best_loc > loc) {
               // When passing loc, don't exceed our current distance from loc.
-              start = std::max(1, 2 * loc - best_loc);
+              start = std::max(1, 2 * loc - (int)best_loc);
             } else {
               // Already passed loc, downhill from here on in.
               break;
